+++ src/Components/CParticleSystem.h
@@ -9,10 +9,10 @@ using namespace std;
 class CParticleSystem : public Component{
 public:
     CParticleSystem();
-    CParticleSystem(unsigned long _nParticles, glm::vec3 _velocity, string _texture, int _width, int _height, int _spawnDelay, int _particlesToSpawn, int _lifeSpan, glm::vec3 _offset, glm::vec3 _orientation, float _radious, std::uint_fast8_t _flags, bool _loop, bool _started);
+    CParticleSystem(uint32_t _nParticles, glm::vec3 _velocity, string _texture, int _width, int _height, int _spawnDelay, int _particlesToSpawn, int _lifeSpan, glm::vec3 _offset, glm::vec3 _orientation, float _radious, std::uint_fast8_t _flags, bool _loop, bool _started);
     ~CParticleSystem() = default;
 
-    unsigned long nParticles {0}; // Numero de particulas
+    uint32_t nParticles {0}; // Numero de particulas
     glm::vec3 velocity;           // Velocidad de las particulas
     string texture;               // Textura de la particula
     int width {50};               // Anchura de la particula
